rtc: Respond to file request with file contents
We no longer store the entire file in memory, just a File object. When we
get a request for a file, we need to read it from disk.

For now we only return the first 100 bytes or so rather than implementing a
reliable transport over a UDP-like data channel. It's better to just wait
until Chrome and other browsers support {reliable: true} for data channels.